Standard wallet functionality - derive keys, gen addrs, sign txs
